@@ -6,9 +6,8 @@
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "format": "prettier . --write",
-    "prepare": "husky"
+    "lint": "eslint .",
+    "format": "prettier . --write"
   },
   "dependencies": {
     "@radix-ui/react-avatar": "^1.1.2",
@@ -23,35 +22,43 @@
     "clsx": "^2.1.1",
     "crypto": "^1.0.1",
     "dotenv": "^16.4.7",
+    "drizzle-orm": "^0.39.3",
     "form-data": "^4.0.1",
     "framer-motion": "^12.0.6",
     "lemonsqueezy": "^0.0.1-claimed",
     "lemonsqueezy.ts": "^0.1.8",
     "lucide-react": "^0.474.0",
     "mailgun.js": "^11.1.0",
-    "next": "15.1.9",
+    "next": "16.0.9",
     "next-auth": "^4.24.11",
-    "nextra": "^4.2.1",
-    "react": "^19.0.1",
-    "react-dom": "^19.0.1",
+    "pg": "^8.13.3",
+    "react": "19.2.0",
+    "react-dom": "19.2.0",
     "react-icons": "^5.4.0",
     "stripe": "^17.6.0",
     "tailwind-merge": "^3.0.1",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
     "@tailwindcss/typography": "^0.5.16",
     "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@types/pg": "^8.11.11",
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
+    "drizzle-kit": "^0.30.4",
     "eslint": "^9",
-    "eslint-config-next": "15.1.6",
-    "husky": "^9.1.7",
+    "eslint-config-next": "16.0.9",
     "postcss": "^8",
     "prettier": "^3.4.2",
     "tailwindcss": "^3.4.1",
+    "tsx": "^4.19.3",
     "typescript": "^5"
   },
-  "packageManager": "pnpm@9.15.3+sha512.1f79bc245a66eb0b07c5d4d83131240774642caaa86ef7d0434ab47c0d16f66b04e21e0c086eb61e62c77efc4d7f7ec071afad3796af64892fae66509173893a"
+  "packageManager": "pnpm@9.15.3+sha512.1f79bc245a66eb0b07c5d4d83131240774642caaa86ef7d0434ab47c0d16f66b04e21e0c086eb61e62c77efc4d7f7ec071afad3796af64892fae66509173893a",
+  "pnpm": {
+    "overrides": {
+      "@types/react": "19.2.2",
+      "@types/react-dom": "19.2.2"
+    }
+  }
 }
