@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import RegisterForm from "@/components/register-form";
+import { RegisterForm } from "@/components/RegisterForm";
 
 export const metadata: Metadata = {
   title: "Register",
