@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Bricolage_Grotesque } from "next/font/google";
 import "./globals.css";
 import type React from "react"; // Import React
+import { cn } from "@/lib/utils";
 
 const bricolageGrotesque = Bricolage_Grotesque({
   subsets: ["latin"],
@@ -32,8 +33,10 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" className={bricolageGrotesque.className}>
-      <body className="antialiased">{children}</body>
+    <html lang="en">
+      <body className={cn(bricolageGrotesque.className, "antialiased")}>
+        {children}
+      </body>
     </html>
   );
 }
