@@ -102,7 +102,7 @@ export default function LoginForm() {
           </Button>
         </div>
         <p className="text-center text-sm">
-          Don't have an account?{" "}
+          Don&apos;t have an account?{" "}
           <Link href="/auth/register" className="text-blue-500 hover:underline">
             Register
           </Link>
