@@ -53,7 +53,7 @@ export default function RegisterForm() {
       </form>
       <p className="text-center text-sm">
         Already have an account?{" "}
-        <Link href="/auth/login" className="text-blue-500 hover:underline">
+        <Link href="/login" className="text-blue-500 hover:underline">
           Login
         </Link>
       </p>
